Gestor_Neto_Contabilidade-main/requirements.txt
requests>=2.31,<3
python-dateutil>=2.8.2,<3
python-dotenv>=1.0,<2
Flask>=3.0,<4
flask-cors>=4.0,<5
SQLAlchemy>=2.0,<3
alembic>=1.12,<2
fastapi>=0.104,<1
uvicorn[standard]>=0.24,<1
